+++ tests/cms/views/view_config.py
@@ -384,6 +384,7 @@ VIEWS = [
                     "longitude": 1,
                     "latitude": 1,
                     "timezone": "Europe/Berlin",
+                    "deepl_renewal_month": 6,
                 },
             ),
         ],