+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 UNRELEASED
 ----------
 
+* [ [#2003](https://github.com/digitalfabrik/integreat-cms/issues/2003) ] Fix HIX widget
+
 
 2023.1.0
 --------